Project setup change

+++ db/migrate/20150131200439_create_participates.rb
@@ -1,6 +1,6 @@
 class CreateParticipates < ActiveRecord::Migration
   def change
-    drop_table :participates
+    drop_table :participates if ActiveRecord::Base.connection.table_exists? :participates
     
     create_table :participates do |t|
       t.belongs_to :user, index: true
